Put Helix manager out of resource manager
